[ADD] crm: Display 'Expected Revenue' currency on quick create form
Purpose
=======

Expected revenue field on pop up have no any indication of currency due to that people enter a dollar sign in the field. Eg: $1234

then an error appears when trying to save saying that the field is invalid but the user doesn’t know why. If the currency symbol was displayed it would solve the problem.

Specification
=============

'Expected Revenue' field on wizard should be monetary in order to display the currency symbol.