[FIX] website_crm_partner_assign: get the lead to forward at save
When forwarding a list of leads to a partner, the lead_id was not send.
As the field `lead_id` is readonly, its value wasn't sent by the client
on save click.

This behaviour of not sending readonly fields in x2many has been correctly
implemented with the new views but some views (like this one) was basing
their behaviour on this breach.

See rev. odoo@4bee515 for more details.

fix-feedback-saas16